chore: refactor folder structure and add prettier
