+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+  "printWidth": 120,
+  "trailingComma": "es5",
+  "bracketSpacing": false,
+  "singleQuote": true,
+  "jsxBracketSameLine": false,
+  "tabWidth": 2
+}
