+++ package.json
@@ -11,6 +11,7 @@
   "license": "",
   "devDependencies": {
     "@figma/plugin-typings": "^1.38.0",
+    "prettier": "^2.5.1",
     "ts-loader": "^9.2.6",
     "typescript": "^4.5.2",
     "webpack": "^5.65.0",
