+++ src/index.ts
@@ -1,9 +1,13 @@
-import { createSID, sidPrefixList } from "./sid";
+// Helpers
+import {generateText} from './helpers';
+
+// Data types
+import {createSID, sidPrefixList} from './data-types/sid';
 
 // Listen for inputs
-figma.parameters.on("input", ({ key, query, result }: ParameterInputEvent) => {
+figma.parameters.on('input', ({key, query, result}: ParameterInputEvent) => {
   switch (key) {
-    case "prefix":
+    case 'prefix':
       result.setSuggestions(sidPrefixList.filter((s) => s.includes(query)));
       break;
     default:
@@ -12,11 +16,11 @@ figma.parameters.on("input", ({ key, query, result }: ParameterInputEvent) => {
 });
 
 // Run plugin
-figma.on("run", async ({ command, parameters }: RunEvent) => {
+figma.on('run', async ({command, parameters}: RunEvent) => {
   // Run function based on command
   switch (command) {
-    case "sid":
-      createSID(parameters.prefix);
+    case 'sid':
+      generateText(createSID(parameters.prefix));
       setTimeout(figma.closePlugin, 500);
       break;
     default:
